added logic to check if coffee order contains correct amount

+++ src/main/scala/coffeeMachine/model/Order.scala
@@ -1,5 +1,5 @@
 package coffeeMachine.model
 
-case class Order(order: String, amount: Float) {
+case class Order(order: String, amount: BigDecimal) {
 
 }
